+++ src/pptx/oxml/ns.py
@@ -2,19 +2,20 @@
 
 from __future__ import annotations
 
-
 # -- Maps namespace prefix to namespace name for all known PowerPoint XML namespaces --
 _nsmap = {
     "a": "http://schemas.openxmlformats.org/drawingml/2006/main",
     "c": "http://schemas.openxmlformats.org/drawingml/2006/chart",
     "cp": "http://schemas.openxmlformats.org/package/2006/metadata/core-properties",
     "ct": "http://schemas.openxmlformats.org/package/2006/content-types",
+    "cx": "http://schemas.microsoft.com/office/drawing/2014/chartex",
     "dc": "http://purl.org/dc/elements/1.1/",
     "dcmitype": "http://purl.org/dc/dcmitype/",
     "dcterms": "http://purl.org/dc/terms/",
     "ep": "http://schemas.openxmlformats.org/officeDocument/2006/extended-properties",
     "i": "http://schemas.openxmlformats.org/officeDocument/2006/relationships/image",
     "m": "http://schemas.openxmlformats.org/officeDocument/2006/math",
+    "mc": "http://schemas.openxmlformats.org/markup-compatibility/2006",
     "mo": "http://schemas.microsoft.com/office/mac/office/2008/main",
     "mv": "urn:schemas-microsoft-com:mac:vml",
     "o": "urn:schemas-microsoft-com:office:office",
@@ -25,7 +26,6 @@ _nsmap = {
     "r": "http://schemas.openxmlformats.org/officeDocument/2006/relationships",
     "sl": "http://schemas.openxmlformats.org/officeDocument/2006/relationships/slideLayout",
     "v": "urn:schemas-microsoft-com:vml",
-    "ve": "http://schemas.openxmlformats.org/markup-compatibility/2006",
     "w": "http://schemas.openxmlformats.org/wordprocessingml/2006/main",
     "w10": "urn:schemas-microsoft-com:office:word",
     "wne": "http://schemas.microsoft.com/office/word/2006/wordml",